+++ lib/slack/block_kit/element/timepicker.rb
@@ -15,10 +15,11 @@ module Slack
 
         TYPE = 'timepicker'
 
-        def initialize(action_id:, placeholder: nil, initial: nil, emoji: nil)
+        def initialize(action_id:, placeholder: nil, initial: nil, emoji: nil, focus_on_load: nil)
           @placeholder = placeholder_text(placeholder, emoji) if placeholder
           @initial_time = initial
           @action_id = action_id
+          @focus_on_load = focus_on_load
 
           yield(self) if block_given?
         end
@@ -29,6 +30,7 @@ module Slack
             action_id: @action_id,
             placeholder: @placeholder&.as_json,
             initial_time: @initial_time,
+            focus_on_load: @focus_on_load,
             confirm: confirm&.as_json
           }.compact
         end