+++ spec/lib/slack/block_kit/element/multi_static_select_spec.rb
@@ -85,6 +85,17 @@ RSpec.describe Slack::BlockKit::Element::MultiStaticSelect do
     ]
   end
 
+  describe '.initialize' do
+    it 'yields self' do
+      yielded = nil
+      new_instance = described_class.new(**params) do |static_select|
+        yielded = static_select
+      end
+
+      expect(new_instance).to be(yielded)
+    end
+  end
+
   describe '#as_json' do
     subject(:as_json) { instance.as_json }
 
@@ -252,6 +263,20 @@ RSpec.describe Slack::BlockKit::Element::MultiStaticSelect do
             initial_options: expected_initial_options
           )
         )
+      end
+    end
+
+    context 'with focus_on_load' do
+      let(:params) do
+        {
+          placeholder: placeholder_text,
+          action_id: action_id,
+          focus_on_load: true
+        }
+      end
+
+      it 'correctly serializes' do
+        expect(as_json).to eq(expected_json.merge(focus_on_load: true))
       end
     end
   end