+++ spec/lib/slack/block_kit/element/multi_users_select_spec.rb
@@ -13,6 +13,17 @@ RSpec.describe Slack::BlockKit::Element::MultiUsersSelect do
     }
   end
 
+  describe '.initialize' do
+    it 'yields self' do
+      yielded = nil
+      new_instance = described_class.new(**params) do |users_select|
+        yielded = users_select
+      end
+
+      expect(new_instance).to be(yielded)
+    end
+  end
+
   describe '#as_json' do
     subject(:as_json) { instance.as_json }
 
@@ -128,6 +139,32 @@ RSpec.describe Slack::BlockKit::Element::MultiUsersSelect do
             text: placeholder_text
           },
           action_id: action_id
+        }
+      end
+
+      it 'correctly serializes' do
+        expect(as_json).to eq(expected_json)
+      end
+    end
+
+    context 'with focus_on_load' do
+      let(:params) do
+        {
+          placeholder: placeholder_text,
+          action_id: action_id,
+          focus_on_load: true
+        }
+      end
+
+      let(:expected_json) do
+        {
+          type: 'multi_users_select',
+          placeholder: {
+            type: 'plain_text',
+            text: placeholder_text
+          },
+          action_id: action_id,
+          focus_on_load: true
         }
       end